+++ tests/profiles/googlefonts_test.py
@@ -2705,6 +2705,10 @@ def test_check_fontdata_namecheck():
     check = CheckTester(googlefonts_profile,
                         "com.google.fonts/check/fontdata_namecheck")
 
+    TIMEOUT_MSG = ("Sometimes namecheck.fontdata.com times out"
+                   " and we don't want to stop running all the other"
+                   " code tests. Unless you touched this portion of"
+                   " the code, it is generaly safe to ignore this glitch.")
     # We dont FAIL because this is meant as a merely informative check
     # There may be frequent cases when fonts are being updated and thus
     # already have a public family name registered on the
@@ -2712,13 +2716,15 @@ def test_check_fontdata_namecheck():
     font = TEST_FILE("cabin/Cabin-Regular.ttf")
     assert_results_contain(check(font),
                            INFO, 'name-collision',
-                           'with an already used name...')
+                           'with an already used name...',
+                           ignore_error=TIMEOUT_MSG)
 
     # Here we know that FamilySans has not been (and will not be)
     # registered as a real family.
     font = TEST_FILE("familysans/FamilySans-Regular.ttf")
     assert_PASS(check(font),
-                'with a unique family name...')
+                'with a unique family name...',
+                ignore_error=TIMEOUT_MSG)
 
 
 def test_check_fontv():