+++ Lib/fontbakery/profiles/googlefonts.py
@@ -83,6 +83,7 @@ FAMILY_CHECKS = [
     'com.google.fonts/check/family/has_license',
     'com.google.fonts/check/family/control_chars',
     'com.google.fonts/check/family/tnum_horizontal_metrics',
+    'com.google.fonts/check/family/italics_have_roman_counterparts',
 ]
 
 NAME_TABLE_CHECKS = [
@@ -4305,6 +4306,60 @@ def com_google_fonts_check_family_control_chars(ttFonts):
                       f"{msg_unacceptable}")
     else:
         yield PASS, ("Unacceptable control characters were not identified.")
+
+
+@check(
+    id = 'com.google.fonts/check/family/italics_have_roman_counterparts',
+    rationale = """
+        For each font family on Google Fonts, every Italic style must have a Roman sibling.
+
+        This kind of problem was first observed at [1] where the Bold style was missing but BoldItalic was included.
+
+        [1] https://github.com/google/fonts/pull/1482
+    """,
+    proposal = 'https://github.com/googlefonts/fontbakery/issues/1733'
+)
+def com_google_fonts_check_family_italics_have_roman_counterparts(fonts):
+    """Ensure Italic styles have Roman counterparts."""
+
+    italics = [f for f in fonts if 'Italic' in f]
+    missing_roman = []
+    for italic in italics:
+        if '-' not in os.path.basename(italic) \
+            or len(os.path.basename(italic).split('-')[-1].split('.')) != 2:
+            yield WARN,\
+                  Message('bad-filename',
+                          f"Filename seems to be incorrect: '{italic}'")
+
+        style = os.path.basename(italic).split('-')[-1].split('.')[0]
+        is_varfont = '[' in style
+
+        # to remove the axes syntax used on variable-font filenames:
+        if is_varfont:
+            style = style.split('[')[0]
+
+        if style == 'Italic':
+            if is_varfont:
+                # "Familyname-Italic[wght,wdth].ttf" => "Familyname[wght,wdth].ttf"
+                roman_counterpart = italic.replace('-Italic', '')
+            else:
+                # "Familyname-Italic.ttf" => "Familyname-Regular.ttf"
+                roman_counterpart = italic.replace('Italic', 'Regular')
+        else:
+            # "Familyname-BoldItalic[wght,wdth].ttf" => "Familyname-Bold[wght,wdth].ttf"
+            roman_counterpart = italic.replace('Italic', '')
+
+        if roman_counterpart not in fonts:
+            missing_roman.append(italic)
+
+    if missing_roman:
+        from fontbakery.utils import pretty_print_list
+        missing_roman = pretty_print_list(missing_roman)
+        yield FAIL,\
+              Message('missing-roman',
+                      f"Italics missing a Roman counterpart: {missing_roman}")
+    else:
+        yield PASS, "OK"
 
 
 @check(